requirements.txt
pandas==2.3.1
langchain==0.3.26
langgraph==0.5.2
langsmith==0.4.5
langchain-openai==0.3.27
langchain-google-genai==2.1.7
pyarrow==20.0.0
tabulate==0.9.0
fastparquet==2024.11.0
matplotlib==3.10.3
litellm==1.74.2
duckdb==1.3.2
python-dotenv==1.1.1
pydantic==2.11.7